ok-gesture-poc/requirements.txt
# OK Sign Gesture Detection - Requirements

# Core ML/DL
ultralytics>=8.0.0
sahi>=0.11.0
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0
openvino>=2023.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0

# Tracking
boxmot>=10.0.0
lap>=0.4.0
scipy>=1.10.0

# Utilities
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
python-dateutil>=2.8.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
black>=23.0.0

# Data annotation (optional - install if needed)
# label-studio>=1.0.0
